requirements.txt
Flask==3.1.*
flask-login==0.6.*
bcrypt==4.*
requests==2.*
gunicorn==23.*
python-dotenv==1.*